Initial upload of the main code
DistKNN.jl contains the main computation code for both CPU and GPU. It also covers the alternative algorithm implementation of the calculation functions with FLANN and ParallelNeighbors.jl

Test_DistKNN.jl contains the main testing code for all the algorithms, as well as it produces the comparison graphs between them automatically.

Test_DistKNN_HPC.jl same as the above, although configured for HPC environments like in the Aristotle computing array. 

read_file.jl is an assistive set of functions for reading .fvec as well as binary files used as input for the algorithms.